talkpro-ai-services/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
aiohttp
gigachat